+++ NNGuiProgram/NeuralNetGuiProgram/customnn.h
@@ -24,7 +24,7 @@ public:
                          double memoryRangeToSmoothingFactor,int activationFunctionType);
     void BackPropogation(const vector<double> &TargetVals, double learnToEta,
                                    double momentumToAlpha, double memoryRangeToSmoothingFactor,
-                                   int activationFunctionType,const vector<int> Outputs,const vector<int> Inputs);
+                                   int activationFunctionType, const vector<double> Outputs, const vector<double> Inputs);
     //since GetResults doesnt actually do anything and just reads
     //the results and spits them back so it is a const function
     //note that the const is not neccessary
